create new test_xgboost.py and move tests for check_objective_supported in
